chore: move scripts to right directory
