+++ scripts/log_telemetry_stats.py
